refactor(mobile): split compose bar widgets
