+++ mobile/lib/features/channels/compose_bar.dart
@@ -16,6 +16,12 @@ import 'channel_management_provider.dart';
 import 'channels_provider.dart';
 import 'emoji_picker.dart';
 
+part 'compose_bar/helpers.dart';
+part 'compose_bar/suggestions.dart';
+part 'compose_bar/formatting_toolbar.dart';
+part 'compose_bar/attachments.dart';
+part 'compose_bar/send_button.dart';
+
 /// Rich compose bar with @mention autocomplete, emoji picker, and a markdown
 /// formatting toolbar. Used in both channel and thread views — the caller
 /// provides an [onSend] callback that handles actual message submission.
@@ -501,610 +507,4 @@ class ComposeBar extends HookConsumerWidget {
       ],
     );
   }
-}
-
-const _typingThrottleMs = 3000;
-
-/// Walk backward from [cursor] looking for [trigger] (e.g. `@` or `#`) at a
-/// word boundary. Returns the index of the trigger character, or `null` if none
-/// is found.
-///
-/// When [stopAtSpace] is `true` the walk stops at both spaces and newlines —
-/// appropriate for `#channel` names which are kebab-case slugs without spaces.
-/// When `false`, only newlines stop the walk, allowing multi-word queries like
-/// `@Alice Smith` to match members with multi-word display names.
-@visibleForTesting
-int? findTrigger(
-  String text,
-  int cursor,
-  String trigger, {
-  bool stopAtSpace = true,
-}) {
-  for (var i = cursor - 1; i >= 0; i--) {
-    final ch = text[i];
-    if (ch == '\n') break;
-    if (stopAtSpace && ch == ' ') break;
-    if (ch == trigger) {
-      if (i == 0 || text[i - 1] == ' ' || text[i - 1] == '\n') {
-        return i;
-      }
-      break;
-    }
-  }
-  return null;
-}
-
-/// Replace the range `[start, cursor)` with [replacement] and move the cursor
-/// to the end of the replacement. Used by both mention and channel insertion.
-@visibleForTesting
-void spliceAndMoveCursor(
-  TextEditingController controller,
-  FocusNode focusNode, {
-  required int start,
-  required String replacement,
-}) {
-  final text = controller.text;
-  final cursor =
-      (controller.selection.isValid
-              ? controller.selection.baseOffset
-              : text.length)
-          .clamp(start, text.length);
-
-  final before = text.substring(0, start);
-  final after = text.substring(cursor);
-  controller.text = '$before$replacement$after';
-  controller.selection = TextSelection.collapsed(
-    offset: start + replacement.length,
-  );
-  focusNode.requestFocus();
-}
-
-/// Insert [trigger] (e.g. `@` or `#`) at the cursor position, prefixed with
-/// a space if needed for word separation. Used by `triggerMention` and
-/// `triggerChannel`.
-void _insertTriggerAtCursor(
-  TextEditingController controller,
-  FocusNode focusNode,
-  String trigger,
-) {
-  final text = controller.text;
-  final cursor = controller.selection.isValid
-      ? controller.selection.baseOffset
-      : text.length;
-  final needsSpace =
-      cursor > 0 && text[cursor - 1] != ' ' && text[cursor - 1] != '\n';
-  final insert = needsSpace ? ' $trigger' : trigger;
-  final before = text.substring(0, cursor);
-  final after = text.substring(cursor);
-  controller.text = '$before$insert$after';
-  controller.selection = TextSelection.collapsed(
-    offset: cursor + insert.length,
-  );
-  focusNode.requestFocus();
-}
-
-/// Send a typing indicator over the WebSocket (fire-and-forget).
-///
-/// Desktop sends these as `["EVENT", signedEvent]` over the WebSocket — not
-/// via HTTP. Ephemeral events like typing indicators are broadcast-only and
-/// the relay doesn't persist them, so the HTTP `/api/events` endpoint may
-/// silently discard them.
-void _sendTypingIndicator(
-  WidgetRef ref, {
-  required String channelId,
-  String? threadHeadId,
-  String? rootId,
-}) {
-  try {
-    final config = ref.read(relayConfigProvider);
-    final nsec = config.nsec;
-    if (nsec == null || nsec.isEmpty) return;
-
-    final privkeyHex = nostr.Nip19.decode(payload: nsec).data;
-    if (privkeyHex.isEmpty) return;
-
-    final tags = <List<String>>[
-      ['h', channelId],
-      if (threadHeadId != null && rootId != null && rootId != threadHeadId) ...[
-        ['e', rootId, '', 'root'],
-        ['e', threadHeadId, '', 'reply'],
-      ] else if (threadHeadId != null)
-        ['e', threadHeadId, '', 'reply'],
-    ];
-
-    final event = nostr.Event.from(
-      kind: EventKind.typingIndicator,
-      content: '',
-      tags: tags,
-      secretKey: privkeyHex,
-      verify: false,
-    );
-
-    // Send directly over WebSocket — fire-and-forget, matching desktop.
-    final session = ref.read(relaySessionProvider.notifier);
-    session.sendRaw(['EVENT', event.toMap()]);
-  } catch (_) {
-    // Fire-and-forget — typing indicator failure is non-fatal.
-  }
-}
-
-List<ChannelMember> _filterMembers(
-  List<ChannelMember> members,
-  String? query,
-  String? currentPubkey,
-  Map<String, UserProfile> userCache,
-) {
-  if (query == null) return const [];
-  final q = query.toLowerCase();
-  return members
-      .where(
-        (m) =>
-            currentPubkey == null ||
-            m.pubkey.toLowerCase() != currentPubkey.toLowerCase(),
-      )
-      .where((m) {
-        if (q.isEmpty) return true;
-        final profile = userCache[m.pubkey.toLowerCase()];
-        final name = (profile?.displayName ?? m.displayName ?? '')
-            .toLowerCase();
-        final firstName = name.split(RegExp(r'\s+')).first;
-        return name.startsWith(q) ||
-            firstName.startsWith(q) ||
-            name.contains(q);
-      })
-      .take(6)
-      .toList();
-}
-
-class _MentionSuggestions extends StatelessWidget {
-  final List<ChannelMember> suggestions;
-  final Map<String, UserProfile> userCache;
-  final String? currentPubkey;
-  final void Function(ChannelMember) onSelect;
-
-  const _MentionSuggestions({
-    required this.suggestions,
-    required this.userCache,
-    required this.currentPubkey,
-    required this.onSelect,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      constraints: const BoxConstraints(maxHeight: 240),
-      clipBehavior: Clip.hardEdge,
-      decoration: BoxDecoration(
-        color: context.colors.surfaceContainerHighest,
-        borderRadius: const BorderRadius.vertical(
-          top: Radius.circular(Radii.dialog),
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: context.colors.shadow.withValues(alpha: 0.08),
-            blurRadius: 8,
-            offset: const Offset(0, -2),
-          ),
-        ],
-      ),
-      child: ListView.separated(
-        shrinkWrap: true,
-        padding: const EdgeInsets.symmetric(vertical: Grid.xxs),
-        itemCount: suggestions.length,
-        separatorBuilder: (_, _) => const SizedBox.shrink(),
-        itemBuilder: (context, index) {
-          final member = suggestions[index];
-          final profile = userCache[member.pubkey.toLowerCase()];
-          final name = profile?.displayName?.trim().isNotEmpty == true
-              ? profile!.displayName!.trim()
-              : member.labelFor(currentPubkey);
-          final avatarUrl = profile?.avatarUrl;
-          final initial =
-              (profile?.displayName?.trim().isNotEmpty == true
-                      ? profile!.displayName!.trim()
-                      : member.pubkey)[0]
-                  .toUpperCase();
-
-          return ListTile(
-            dense: true,
-            visualDensity: VisualDensity.compact,
-            leading: CircleAvatar(
-              radius: 14,
-              backgroundColor: context.colors.primaryContainer,
-              backgroundImage: avatarUrl != null
-                  ? NetworkImage(avatarUrl)
-                  : null,
-              child: avatarUrl == null
-                  ? Text(
-                      initial,
-                      style: context.textTheme.labelSmall?.copyWith(
-                        color: context.colors.onPrimaryContainer,
-                      ),
-                    )
-                  : null,
-            ),
-            title: Text(name, style: context.textTheme.bodyMedium),
-            trailing: member.isBot
-                ? Icon(
-                    LucideIcons.bot,
-                    size: 14,
-                    color: context.colors.onSurfaceVariant,
-                  )
-                : null,
-            onTap: () => onSelect(member),
-          );
-        },
-      ),
-    );
-  }
-}
-
-@visibleForTesting
-List<Channel> filterChannels(List<Channel> channels, String? query) {
-  if (query == null) return const [];
-  final q = query.toLowerCase();
-  return channels
-      .where((c) => c.channelType != 'dm')
-      .where((c) {
-        if (q.isEmpty) return true;
-        return c.name.toLowerCase().contains(q);
-      })
-      .take(8)
-      .toList();
-}
-
-class _ChannelSuggestions extends StatelessWidget {
-  final List<Channel> suggestions;
-  final void Function(Channel) onSelect;
-
-  const _ChannelSuggestions({
-    required this.suggestions,
-    required this.onSelect,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      constraints: const BoxConstraints(maxHeight: 240),
-      clipBehavior: Clip.hardEdge,
-      decoration: BoxDecoration(
-        color: context.colors.surfaceContainerHighest,
-        borderRadius: const BorderRadius.vertical(
-          top: Radius.circular(Radii.dialog),
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: context.colors.shadow.withValues(alpha: 0.08),
-            blurRadius: 8,
-            offset: const Offset(0, -2),
-          ),
-        ],
-      ),
-      child: ListView.separated(
-        shrinkWrap: true,
-        padding: const EdgeInsets.symmetric(vertical: Grid.xxs),
-        itemCount: suggestions.length,
-        separatorBuilder: (_, _) => const SizedBox.shrink(),
-        itemBuilder: (context, index) {
-          final channel = suggestions[index];
-          return ListTile(
-            dense: true,
-            visualDensity: VisualDensity.compact,
-            leading: Icon(
-              channel.isForum ? LucideIcons.messageSquare : LucideIcons.hash,
-              size: 18,
-              color: context.colors.onSurfaceVariant,
-            ),
-            title: Text(
-              '#${channel.name}',
-              style: context.textTheme.bodyMedium,
-            ),
-            trailing: Text(
-              channel.channelType,
-              style: context.textTheme.labelSmall?.copyWith(
-                color: context.colors.onSurfaceVariant,
-              ),
-            ),
-            onTap: () => onSelect(channel),
-          );
-        },
-      ),
-    );
-  }
-}
-
-class _FormattingToolbar extends StatelessWidget {
-  final void Function(String prefix, [String? suffix]) onFormat;
-
-  const _FormattingToolbar({required this.onFormat});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: Grid.half),
-      child: Row(
-        children: [
-          _FormatButton(
-            icon: LucideIcons.bold,
-            tooltip: 'Bold',
-            onTap: () => onFormat('**'),
-          ),
-          _FormatButton(
-            icon: LucideIcons.italic,
-            tooltip: 'Italic',
-            onTap: () => onFormat('_'),
-          ),
-          _FormatButton(
-            icon: LucideIcons.strikethrough,
-            tooltip: 'Strikethrough',
-            onTap: () => onFormat('~~'),
-          ),
-          _FormatButton(
-            icon: LucideIcons.code,
-            tooltip: 'Code',
-            onTap: () => onFormat('`'),
-          ),
-          _FormatButton(
-            icon: LucideIcons.squareCode,
-            tooltip: 'Code block',
-            onTap: () => onFormat('```\n', '\n```'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _FormatButton extends StatelessWidget {
-  final IconData icon;
-  final String tooltip;
-  final VoidCallback onTap;
-
-  const _FormatButton({
-    required this.icon,
-    required this.tooltip,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Tooltip(
-      message: tooltip,
-      child: InkWell(
-        borderRadius: BorderRadius.circular(Radii.sm),
-        onTap: onTap,
-        child: Padding(
-          padding: const EdgeInsets.all(Grid.xxs),
-          child: Icon(icon, size: 18, color: context.colors.onSurfaceVariant),
-        ),
-      ),
-    );
-  }
-}
-
-class _ComposeAction extends StatelessWidget {
-  final IconData icon;
-  final bool active;
-  final VoidCallback onTap;
-
-  const _ComposeAction({
-    required this.icon,
-    this.active = false,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 36,
-      height: 36,
-      child: IconButton(
-        onPressed: onTap,
-        icon: Icon(
-          icon,
-          size: 20,
-          color: active
-              ? context.colors.primary
-              : context.colors.onSurfaceVariant,
-        ),
-        padding: EdgeInsets.zero,
-        visualDensity: VisualDensity.compact,
-      ),
-    );
-  }
-}
-
-@immutable
-class _ComposeDraftPayload {
-  final String content;
-  final List<List<String>> mediaTags;
-
-  const _ComposeDraftPayload({required this.content, required this.mediaTags});
-
-  factory _ComposeDraftPayload.fromDraft({
-    required String text,
-    required List<BlobDescriptor> attachments,
-    required List<CustomEmoji> customEmoji,
-  }) {
-    var content = text;
-    final mediaTags = <List<String>>[];
-    for (final attachment in attachments) {
-      mediaTags.add(attachment.toImetaTag());
-      content += '\n${attachment.toMarkdownImage()}';
-    }
-    mediaTags.addAll(buildCustomEmojiTags(content, customEmoji));
-    return _ComposeDraftPayload(content: content, mediaTags: mediaTags);
-  }
-}
-
-List<BlobDescriptor> _withoutAttachment(
-  List<BlobDescriptor> attachments,
-  String url,
-) {
-  return [
-    for (final attachment in attachments)
-      if (attachment.url != url) attachment,
-  ];
-}
-
-class _AttachmentStrip extends StatelessWidget {
-  final List<BlobDescriptor> attachments;
-  final int uploadingCount;
-  final void Function(String url) onRemove;
-
-  const _AttachmentStrip({
-    required this.attachments,
-    required this.uploadingCount,
-    required this.onRemove,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final thumbWidth = 72.0;
-    final thumbHeight = 72.0;
-
-    return SizedBox(
-      height: thumbHeight,
-      child: ListView.separated(
-        scrollDirection: Axis.horizontal,
-        itemCount: attachments.length + uploadingCount,
-        separatorBuilder: (_, _) => const SizedBox(width: Grid.half),
-        itemBuilder: (context, index) {
-          if (index >= attachments.length) {
-            return Container(
-              width: thumbWidth,
-              decoration: BoxDecoration(
-                color: context.colors.surface,
-                borderRadius: BorderRadius.circular(Radii.md),
-                border: Border.all(color: context.colors.outlineVariant),
-              ),
-              child: const Center(
-                child: CircularProgressIndicator(strokeWidth: 2),
-              ),
-            );
-          }
-
-          final attachment = attachments[index];
-          final isVideo = attachment.type.startsWith('video/');
-          final previewUrl = attachment.thumb ?? attachment.url;
-          return Container(
-            key: ValueKey('compose-attachment:${attachment.url}'),
-            width: thumbWidth,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(Radii.md),
-              border: Border.all(color: context.colors.outlineVariant),
-            ),
-            child: Stack(
-              fit: StackFit.expand,
-              children: [
-                ClipRRect(
-                  borderRadius: BorderRadius.circular(Radii.md),
-                  child: isVideo
-                      ? ColoredBox(
-                          color: Colors.black,
-                          child: Center(
-                            child: Icon(
-                              LucideIcons.video,
-                              color: Colors.white,
-                              size: 24,
-                            ),
-                          ),
-                        )
-                      : Image.network(
-                          previewUrl,
-                          fit: BoxFit.cover,
-                          errorBuilder: (_, _, _) => ColoredBox(
-                            color: context.colors.surface,
-                            child: Icon(
-                              LucideIcons.image,
-                              color: context.colors.onSurfaceVariant,
-                            ),
-                          ),
-                        ),
-                ),
-                Positioned(
-                  top: Grid.quarter,
-                  right: Grid.quarter,
-                  child: SizedBox(
-                    width: 24,
-                    height: 24,
-                    child: IconButton(
-                      onPressed: () => onRemove(attachment.url),
-                      tooltip: 'Remove attachment',
-                      visualDensity: VisualDensity.compact,
-                      style: IconButton.styleFrom(
-                        backgroundColor: context.colors.surface.withValues(
-                          alpha: 0.92,
-                        ),
-                        minimumSize: const Size(24, 24),
-                        maximumSize: const Size(24, 24),
-                        padding: EdgeInsets.zero,
-                        tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                      ),
-                      icon: Icon(
-                        LucideIcons.x,
-                        size: 14,
-                        color: context.colors.onSurface,
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
-
-class _SendButton extends StatelessWidget {
-  final bool isSending;
-  final bool isDisabled;
-  final VoidCallback onTap;
-
-  const _SendButton({
-    required this.isSending,
-    required this.onTap,
-    this.isDisabled = false,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 36,
-      height: 36,
-      child: IconButton(
-        onPressed: (isSending || isDisabled) ? null : onTap,
-        style: IconButton.styleFrom(
-          backgroundColor: context.colors.primary,
-          disabledBackgroundColor: context.colors.primary.withValues(
-            alpha: 0.5,
-          ),
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(Radii.md),
-          ),
-        ),
-        padding: EdgeInsets.zero,
-        icon: isSending
-            ? SizedBox(
-                width: 16,
-                height: 16,
-                child: CircularProgressIndicator(
-                  strokeWidth: 2,
-                  color: context.colors.onPrimary,
-                ),
-              )
-            : Icon(
-                LucideIcons.sendHorizontal,
-                size: 18,
-                color: context.colors.onPrimary,
-              ),
-      ),
-    );
-  }
-}
-
-String _formatUploadError(Object error) {
-  return error.toString().replaceFirst('Exception: ', '');
 }
